Order status is added and is changed overtime using HangFire
order status changes from pending to shipped to delivered. once the user places an order, the order status will be initially pending, after a certain time (for educational purpose time is set to 1 minute) the HnagFire Sceduling method fires in the backgroung and changes the status from pending to shipped, then after that it is changed to delivered.